+++ test/features/item_list/data/datasources/item_remote_data_source_test.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:caramelseed/core/authorization/token_manager.dart';
+import 'package:caramelseed/core/managers/token_manager.dart';
 import 'package:caramelseed/core/commons/common_response.dart';
 import 'package:caramelseed/core/config/env_config.dart';
 import 'package:caramelseed/core/error/exceptions.dart';
@@ -67,55 +67,43 @@ void main() {
   }
 
   group('Get All', () {
-    final CommonResponse<ItemModel> commonResponse = CommonResponse<ItemModel>.fromJson(
-        jsonDecode(fixture('get_all_items.json')) as Map<String, dynamic>);
+    final CommonResponse<ItemModel> commonResponse =
+        CommonResponse<ItemModel>.fromJson(jsonDecode(fixture('get_all_items.json')) as Map<String, dynamic>);
     final List<Item> itemList = commonResponse.data;
-    test(
-      'should return an Item List when the response code is 200 (success).',
-      () async {
-        // arrange
-        setUpMockHttpClientSuccess200();
-        // act
-        final result = await dataSource.getAll();
-        // assert
-        expect(result, equals(itemList));
-      }, skip: true
-    );
+    test('should return an Item List when the response code is 200 (success).', () async {
+      // arrange
+      setUpMockHttpClientSuccess200();
+      // act
+      final result = await dataSource.getAll();
+      // assert
+      expect(result, equals(itemList));
+    }, skip: true);
 
-    test(
-      'should throw a ServerException when the response code is 404 or other.',
-      () async {
-        // arrange
-        setUpMockHttpClientFailure404();
-        // act
-        final call = dataSource.getAll;
-        // assert
-        expect(() => call(), throwsA(const TypeMatcher<ServerException>()));
-      }, skip: true
-    );
+    test('should throw a ServerException when the response code is 404 or other.', () async {
+      // arrange
+      setUpMockHttpClientFailure404();
+      // act
+      final call = dataSource.getAll;
+      // assert
+      expect(() => call(), throwsA(const TypeMatcher<ServerException>()));
+    }, skip: true);
 
-    test(
-      'should throw a ServerException when the token is not present on the storage.',
-      () async {
-        // arrange
-        when(mockTokenManager.saveToken(any)).thenThrow(GetTokenException());
-        // act
-        final call = dataSource.getAll;
-        // assert
-        expect(() => call(), throwsA(const TypeMatcher<ServerException>()));
-      }, skip: true
-    );
+    test('should throw a ServerException when the token is not present on the storage.', () async {
+      // arrange
+      when(mockTokenManager.saveToken(any)).thenThrow(GetTokenException());
+      // act
+      final call = dataSource.getAll;
+      // assert
+      expect(() => call(), throwsA(const TypeMatcher<ServerException>()));
+    }, skip: true);
 
-    test(
-      'should throw a ServerException when the request throw and exception.',
-      () async {
-        // arrange
-        setUpMockHttpClientException();
-        // act
-        final call = dataSource.getAll;
-        // assert
-        expect(() => call(), throwsA(const TypeMatcher<ServerException>()));
-      }, skip: true
-    );
+    test('should throw a ServerException when the request throw and exception.', () async {
+      // arrange
+      setUpMockHttpClientException();
+      // act
+      final call = dataSource.getAll;
+      // assert
+      expect(() => call(), throwsA(const TypeMatcher<ServerException>()));
+    }, skip: true);
   });
 }
